requirements.txt
click==7.1.2
Django==2.2
django-cors-headers==3.5.0
django-picklefield==3.0.1
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.0.4
numpy==1.19.4
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.4.1
